[Refactor] SPA 형식으로 변경
